+++ samples/Crabalidator.Samples.Basic/Address.cs
@@ -1,0 +1,7 @@
+namespace Crabalidator.Samples.Basic
+{
+    public sealed class Address
+    {
+        public string PostalCode { get; set; }
+    }
+}