+++ samples/Crabalidator.Samples.Basic/Customer.cs
@@ -1,0 +1,13 @@
+namespace Crabalidator.Samples.Basic
+{
+    public sealed class Customer
+    {
+        public string Name { get; set; }
+
+        public int Age { get; set; }
+
+        public string Email { get; set; }
+
+        public Address Address { get; set; } = new Address();
+    }
+}